requirements.txt
google-genai
streamlit
tqdm
langchain
langchain-community
langchain-google-genai
faiss-cpu
pypdf